Use a `HashSet` for retaining words
This is an imrpovement on `Vec`, but we could instead limit up front
rather than adding a bunch of tallied words that will be discarded.

The "only" words are a second class feature I wonder about removing?
Alternatively, we could give them a performant path. I think I'd tend to
grep and pipe to a WC if I cared about a specific set of words only.

Maybe do one thing well and remove feature creap? Then again, not we
could handwave performance or actually optimize. TBD.